# dev/transit-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API Transit",
        "description": "Documentation de l'API trans.it générée avec Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/document-params/{id}": {
            "get": {
                "tags": [
                    "ParametreDocument"
                ],
                "summary": "Récupérer un paramètre de document par son ID",
                "operationId": "deb4aa3141b0f8ef65297d8f1b348667",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du paramètre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du paramètre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ID": {
                                            "type": "integer"
                                        },
                                        "Protection_MotDePasse": {
                                            "type": "boolean"
                                        },
                                        "Date_Expiration": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "Date_Update": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "Date_Creation": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paramètre document introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ParametreDocument"
                ],
                "summary": "Mettre à jour les paramètres d'un document",
                "operationId": "45b4a72a1f258ebec93a92fb034b6ab2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du paramètre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Protection_MotDePasse": {
                                        "type": "boolean"
                                    },
                                    "Mot_de_passe": {
                                        "type": "string"
                                    },
                                    "Date_Expiration": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "404": {
                        "description": "Paramètre introuvable"
                    },
                    "500": {
                        "description": "Erreur mise à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ParametreDocument"
                ],
                "summary": "Supprimer les paramètres d'un document",
                "operationId": "567a35528953c863c8e5dbd6abf85f24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du paramètre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimé avec succès"
                    },
                    "404": {
                        "description": "Paramètre introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/document-params": {
            "get": {
                "tags": [
                    "ParametreDocument"
                ],
                "summary": "Récupérer les paramètres de document d'un utilisateur",
                "operationId": "aebe3583e95d3d879ad6efb8b885b966",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paramètre trouvé"
                    },
                    "404": {
                        "description": "Aucun paramètre document trouvé pour cet utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/document-params": {
            "post": {
                "tags": [
                    "ParametreDocument"
                ],
                "summary": "Créer des paramètres pour un document",
                "operationId": "cb79be9be2edf84cb33a79c376100eb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_Document"
                                ],
                                "properties": {
                                    "Id_Document": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "Protection_MotDePasse": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "Mot_de_passe": {
                                        "type": "string",
                                        "example": "Secret123"
                                    },
                                    "Date_Expiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paramètre créé avec succès"
                    },
                    "500": {
                        "description": "Erreur création paramètre"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}